# SafeReturnSwagger.yaml
openapi: 3.0.3
info:
  title: SafeReturn API
  description: |-
    SafeReturn est une application innovante conçue pour offrir une interface centralisée, permettant aux utilisateurs de signaler et de rechercher des animaux perdus et retrouvés. L'application regroupe un ensemble d'annonces de recherche d'animaux disparus sur une seule plateforme afin de faciliter les recherches. SafeReturn intègre des fonctionnalités avancées telles que la géolocalisation, des filtres personnalisés, des notifications instantanées, la gestion des récompenses, ainsi qu'une communication directe entre les propriétaires d'animaux et les chercheurs.
    
    L'API de SafeReturn couvre les fonctionnalités suivantes :
    - **Création et gestion des utilisateurs** : Cette fonctionnalité permet d'ajouter, de modifier, de supprimer et de récupérer les informations relatives aux utilisateurs de l'application..
    - **Création et gestion de déclarations de perte d'animal** : L'API permet l'ajout, la modification, la suppression et la récupération des déclarations concernant les animaux perdus ou retrouvés, ainsi que toutes les informations associées.
    - **Création et gestion de d'abonnement'** : Cette fonctionnalité permet aux utilisateurs de souscrire, de modifier, de résilier et de consulter les informations liées à leurs abonnements.
    - **Authentification et autorisation** : L'API valide les identifiants des utilisateurs et gère les tokens JWT pour sécuriser les accès.
...
    
    
...

...
servers:
...
paths:
...
  /v1/users:
    post:
      tags:
        - users
      summary: Créer un profil utilisateur
      description: Enregistrement d'un utilisateur dans l'application.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        description: Détails de l'utilisateur
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                password:
                  type: string
              example:
                username: "Lionel"
                password: "123456780"
                email: "unemail@gmail.com"
                phone: "+33754871209"

      responses:
        '201':
          description: Profil utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'

        '400':
          description: Données utilisateur invalides

    get:
      tags:
        - users
      summary: Liste des utilisateurs
      description: Récupère l'ensemble des utilisateurs enregistrés. La récupération en mode pagination est disponible grâce au paramètre "page". Si le paramètre page n'est pas renseigné alors la réponse contiendra l'ensemble des utilisateurs, sinon la réponse contiendra un paquet de 50 utilisateurs.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Récupérer les utilisateurs en mode pagination. La numérotation des pages commencent au numéro 0, ainsi pour requêter la première page il faudra mettre ce paramètre à 0.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste des utilisateurs avec pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaginatedUsers'
        '404':
          description: Aucun utilisateur trouvé
        '401':
          description: Utilisateur non vérifié
  /v1/users/{id}:
    get:
      tags:
        - users
      summary: Obtenir un utilisateur
      description: Récupère les informations d'un utilisateur à partir de son ID.
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID de l'utilisateur
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '404':
          description: Utilisateur non trouvé
    delete:
      tags:
        - users
      summary: Supprimer un utilisateur
      description: Supprime un utilisateur à partir de son ID.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID de l'utilisateur à supprimer
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Utilisateur supprimé avec succès
        '404':
          description: Utilisateur non trouvé
        '401':
          description: Non autorisé

    patch:
      tags:
        - users
      summary: Mettre à jour partiellement un utilisateur
      description: Permet de mettre à jour des informations spécifiques d'un utilisateur.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID de l'utilisateur à mettre à jour
          required: true
          schema:
            type: integer
      requestBody:
        description: Champs à mettre à jour dans le profil utilisateur
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                phone:
                  type: string
              example:
                username: "Lionel"
                email: "unemail@gmail.com"
                phone: "+33754871209"
      responses:
        '200':
          description: Utilisateur mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Données invalides
        '404':
          description: Utilisateur non trouvé
  /v1/users/login:
    post:
      tags:
        - users
      summary: Authentifier un utilisateur
      description: Permet aux utilisateurs de se connecter et de recevoir un token d'authentification.
      operationId: login
      requestBody:
        description: Informations de connexion de l'utilisateur
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Échec de l'authentification (identifiants invalides)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthRequest:
      type: object
      properties:
        username:
          type: string
          description: Nom de profil de l'utilisateur
        password:
          type: string
          format: password
          description: Mot de passe de l'utilisateur
      example:
        username: "Lionel112"
        password: "1234567890"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT pour l'accès sécurisé
      example:
        token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
...
    Users:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id unique de l'utilisateur
        username:
          type: string
          description: Nom de profil de l'utilisateur
        password:
          type: string
          description: mot de passe de l'utilisateur
        email:
          type: string
          description: Email de l'utilisateur
        phone:
          type: string
          description: Numéro de téléphone de l'utilisateur. Doit respecter le format suivant, +33XXXXXXXXX en adaptant +33 avec le numéro associé à votre pays
      example:
        id: 40
        username: "Lionel"
        password: "123456780"
        email: "unemail@gmail.com"
        phone: "+33754871209"
...
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Users'
        totalItems:
          type: integer
          description: Nombre total d'utilisateurs
        totalPages:
          type: integer
          description: Nombre total de pages
        currentPage:
          type: integer
          description: Numéro de la page actuelle
      example:
        items:
          - id: 40
            username: "Lionel"
            email: "unemail@gmail.com"
            phone: "+33754871209"
        totalItems: 1
        totalPages: 1
        currentPage: 1
...